+++ .gitignore
@@ -2,4 +2,5 @@
 build/
 vcpkg_installed/
 .cache/
+# config
 .vscode/
